refactor(racingcar): validator 세부화
