+++ src/main/kotlin/racingcar/domain/validator/NumberValidator.kt
@@ -1,0 +1,7 @@
+package racingcar.domain.validator
+
+import java.util.regex.Pattern
+
+object NumberValidator {
+    fun isNumber(value: String): Boolean = Pattern.matches("^[0-9]+$", value)
+}
