+++ src/test/kotlin/racingcar/domain/InputValidatorTest.kt
@@ -3,21 +3,14 @@ package racingcar.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.domain.validator.InputValidator
 
-class InputValueValidatorTest {
+class InputValidatorTest {
     @ValueSource(strings = ["", "jinini"])
     @ParameterizedTest
     fun `자동차 이름이 짧거나 긴 경우 오류 발생`(input: String) {
         assertThat(
-            InputValueValidator.isNameLengthInRange(input),
-        ).isFalse
-    }
-
-    @ValueSource(strings = ["", "a", "0d9", "-1"])
-    @ParameterizedTest
-    fun `숫자가 올바르지 않은 경우 오류 발생`(input: String) {
-        assertThat(
-            InputValueValidator.isNumber(input),
+            InputValidator.isNameLengthInRange(input),
         ).isFalse
     }
 }
